// codebases/trello.com/app/src/components/TemplatePicker/types.ts
/* eslint-disable @trello/disallow-filenames */
export enum Category {
  Popular = 'popular',
  Business = 'business',
  Design = 'design',
  Education = 'education',
  Engineering = 'engineering',
  Marketing = 'marketing',
  HumanResources = 'operations-hr',
  Operations = 'operations',
  SalesCRM = 'sales-crm',
}